Add files via upload
Updated .sql files

+++ DELETE Dummy Script.sql
@@ -1,9 +1,9 @@
-DELETE FROM clean_squat.user;
-DELETE FROM clean_squat.building;
-DELETE FROM clean_squat.session;
-DELETE FROM clean_squat.restroom;
-DELETE FROM clean_squat.admin_edit_building;
-DELETE FROM clean_squat.admin_edit_restroom;
-DELETE FROM clean_squat.manipulate;
-DELETE FROM clean_squat.rating;
-DELETE FROM clean_squat.issuereport;
+DELETE FROM clean_squat.session WHERE SessionID;
+DELETE FROM clean_squat.rating WHERE RatingID;
+DELETE FROM clean_squat.admin_edit_restroom WHERE RestroomID;
+DELETE FROM clean_squat.admin_edit_building WHERE BuildingID;
+DELETE FROM clean_squat.manipulate WHERE IssueID;
+DELETE FROM clean_squat.issuereport WHERE IssueID;
+DELETE FROM clean_squat.restroom WHERE RestroomID;
+DELETE FROM clean_squat.building WHERE BuildingID;
+DELETE FROM clean_squat.user WHERE UserID;
